requirements.txt
Django==4.1.7
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
social-auth-app-django==5.0.0
django-filter==22.1
marshmallow-dataclass==8.5.11

pytest==7.2.1
marshmallow==3.19.0
requests~=2.28.1
environ~=1.0
python-dotenv~=0.21.1